requirements.txt
flask==3.0.0
python-dotenv==1.0.0
crewai==0.121.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
openai>=1.68.2
python-dateutil==2.8.2
gunicorn==21.2.0
requests==2.31.0 